Rename course unit student row widget to course unit student tile for consistency
